### 2026-06-04 #10647 @editorialbot
[PRE REVIEW]: TEGAT: A Lightweight and Reusable Gateway for Scientific Sensor Networks
**Submitting author:** @patrickjaigner (<a href="http://orcid.org/0000-0002-1530-415X">Patrick Aigner</a>)
**Repository:** https://github.com/tum-esm/TEGAT
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** v1.0.1
