### 2026-06-10 c88f991
reduce number of words
