+++ joss/paper.md
@@ -62,11 +62,11 @@ and independently deployed by TEGAT, thus separating infrastructure and applicat
 # Statement of need
 
 Distributed sensor networks are a critical tool in scientific research and widely used across disciplines, enabling 
-long-term, continuous sensor measurements. While they vary in the sensor hardware, data acquisition and processing 
-protocols used, as well as in the number of deployed devices, such networks often face common infrastructure challenges:
+long-term, continuous sensor measurements. While they vary in size and in the type of sensor hardware and data processing
+protocols used, such networks often face common infrastructure challenges:
 Physical access to sensor devices is often limited or costly. This can be additionally exacerbated by challenges in 
 scaling networks to large numbers of deployed sensors. Sensor networks often need to provide continuous
-measurements while operating unattended for extended periods of time. At the same time, network connectivity and system 
+measurements while operating unattended for extended periods of time, all while network connectivity and system 
 power can be intermittent or unreliable.
 Although these challenges vary across deployments, they translate into a common set of infrastructure requirements:
 
@@ -78,10 +78,9 @@ Although these challenges vary across deployments, they translate into a common 
 - Seamless addition and removal of devices from the network
 
 TEGAT addresses these infrastructure requirements to significantly reduce the engineering overhead associated 
-with deploying and maintaining sensor networks. This enables network operators to focus on application-specific logic 
-while building on top of field-tested software. TEGAT leverages the ThingsBoard IoT platform (@ThingsBoard), a 
-robust open source software, which was chosen for its maturity (10+yrs in development) and scalability, supporting large 
-numbers of sensor devices.
+with deploying and maintaining sensor networks. This enables network operators to focus on application-specific instead. 
+TEGAT leverages the ThingsBoard IoT platform (@ThingsBoard), a robust open source software, which was chosen for its 
+maturity (10+yrs in development) and scalability, supporting large numbers of sensor devices.
 The flexible design of TEGAT, combined with the ThingsBoard IoT platform, enables users to configure customized 
 sensor networks, seamlessly integrate additional sensors into existing networks, as well as making it possible to reuse 
 infrastructure across multiple research projects.
@@ -89,17 +88,15 @@ infrastructure across multiple research projects.
 # State of the Field
 Other existing solutions already cover a variety of the features provided by the combination of TEGAT and ThingsBoard,
 though there are different tradeoffs and limitations to consider with each approach:
-Some solutions implement a semi-distributed architecture, where in addition to a central backend server, an on-site 
-central gateway server collects data from multiple connected sensor devices. Examples include thin-edge.io (@thin-edge) and ThingsBoard's
-own product Thingsboard Edge (@thingsboard-edge). These architectures benefit from sensor network layouts where many sensor devices share a local
-network, such as in large factory or office buildings, but face limitations when sensors are deployed individually in
-remote locations and don't form local networks. Furthermore, since the gateway servers are not designed to be co-deployed
-on the sensor devices themselves, they lack software update or remote management capabilities.
+Some solutions are designed around an on-site centralized "edge" server which collects data from multiple connected 
+sensor devices. Examples include thin-edge.io (@thin-edge) and ThingsBoard's own product Thingsboard Edge (@thingsboard-edge). 
+These architectures benefit from sensor network layouts where many sensor devices share a local network, such as large 
+factory or office buildings, but face limitations when sensors are deployed individually in remote locations.
 Some subset of our features can be covered with commercial solutions: For example, Amazon's AWS IoT (@AWS-IoT) and Microsoft
 Azure IoT Edge (@Azure-IoT-Edge) products are IoT cloud-platforms similar to ThingsBoard, and Balena Cloud (@BalenaCloud) offers reliable device 
[…]
