### 2026-06-10 f61bb23
further reduce nr of words
