### 2026-06-10 ba2a018
collaborative word count reduction lars/patrick
