### 2026-06-10 192f8f2
shorten tags and figure captions
