### 2026-06-10 2da4264
shorten SoF
