### 2026-06-10 a0431b6
further shortening
