+++ joss/paper.md
@@ -37,12 +37,10 @@ bibliography: paper.bib
 
 # Summary
 
-TEGAT (Telemetry Edge Gateway) is a lightweight, general-purpose open source software built for managing and monitoring networks of IoT 
-devices. 
+TEGAT (Telemetry Edge Gateway) is a lightweight, general-purpose open source software built for managing and monitoring networks of IoT devices. 
 It is built to integrate with the ThingsBoard platform and was designed specifically for operating distributed 
 sensor networks for scientific research.
-Originally developed for and validated in the ACROPOLIS urban CO2 sensor network, it is network-agnostic and can be used with a 
-wide range of sensor hardware.
+Originally developed for and validated in the ACROPOLIS urban CO2 sensor network, it is network-agnostic and can be used with a wide range of sensor hardware.
 TEGAT offers a stable and reusable architectural baseline for distributed sensor networks, enabling users to 
 focus on application-specific logic while relying on a field-tested software solution for communication, data 
 persistence, and remote management of sensor devices.
@@ -62,7 +60,7 @@ measurements while operating unattended for extended periods of time, all while 
 power can be intermittent or unreliable.
 Although these challenges vary across deployments, they translate into a common set of infrastructure requirements:
 
-- Reliable bidirectional communication with a central platform
+- Reliable bidirectional communication
 - Local buffering of data during network outages
 - Remote configuration and maintenance capabilities
 - Safe remotely initiated software updates
@@ -90,8 +88,7 @@ management and software updates of IoT device fleets similar to TEGAT's OTA and 
 building on top of such products are dependent on their future pricing and availability, and require
 continuous funding.
 Finally, a combination of open source solutions can offer a similar feature set: Examples are the Eclipse Foundation's Kura (@EF-kura) 
-and Kapua (@EF-kapua) projects, as well as the Linux Foundation's Fledge (@fledge) and Kube Edge (@kubeedge) projects. In both cases, these unfortunately lack
-in some aspects we consider important, such as data visualization dashboards and software maturity. Finally, the Ivy project (@Makowski2025) lacks separation between application and infrastructure logic, making software updates brittle. For example, 
+and Kapua (@EF-kapua) projects, as well as the Linux Foundation's Fledge (@fledge) and Kube Edge (@kubeedge) projects. In both cases, these unfortunately lack data visualization dashboards and software maturity. Finally, the Ivy project (@Makowski2025) doesn't separate application and infrastructure logic, making software updates brittle. For example, 
 any crashes not covered by the test suite may result in permanent downtime requiring on-site fixes.
 
 # Software Architecture
@@ -119,7 +116,7 @@ Decoupling TEGAT from the Controller software ensures that corrective actions, s
 reverting to a stable software version or adjusting configurations, can be performed remotely without risking system 
 connectivity or requiring on-site intervention.
 In case a newly deployed Controller Software version fails to start or contains errors, TEGAT remains operational
-and continues to communicate with the ThingsBoard IoT platform. 
+and continues to communicate with the ThingsBoard platform. 
 
 ![Software architecture for on-device (green) and off-premise (blue) components. Purple boxes show the three main architecture components TEGAT (1), Controller Software (2), and the ThingsBoard IoT Platform (3). Dashed boxes show local files. \label{fig:architecture}](figures/figure1.png)
 
@@ -133,7 +130,7 @@ This design ensures operational reliability and efficiency.
 TEGAT receives telemetry data from the Controller Software via a local sqlite database (@SQLite), which is used to
 buffer messages between the two software components for additional fault tolerance. TEGAT then forwards the 
 telemetry data to the ThingsBoard platform via MQTT, and stores a copy of the data in a local database for additional 
[…]
