### 2026-06-10 9c298f8
joss joint writing session: reduce number of words (#26)
* reduce number of words

* further reduce nr of words

* collaborative word count reduction lars/patrick

* shorten tags and figure captions

* shorten SoF

* further shortening
