+++ joss/paper.md
@@ -4,22 +4,14 @@ tags:
   - Gateway
   - Thingsboard
   - IoT
-  - Internet of Things
   - Python
   - ACROPOLIS
   - MQTT
   - Sensor network
-  - Wireless sensor network
   - Sensor management
   - Real-time Measurement
-  - Reproducible research
-  - Time series data
-  - Environmental sensing
-  - Software Updates
-  - Remote Configuration
-  - Remote Procedure Call
-  - Data persistance
-  - Distributed systems
+  - Data persistence
+
 authors:
   - name: Lars Frölich
     orcid: 0009-0000-1579-7727
@@ -45,32 +37,30 @@ bibliography: paper.bib
 
 # Summary
 
-TEGAT (Telemetry Edge Gateway) is a lightweight, general-purpose open source software built for managing and monitoring networks of IoT 
-devices. 
+TEGAT (Telemetry Edge Gateway) is a lightweight, general-purpose open source software built for managing and monitoring networks of IoT devices. 
 It is built to integrate with the ThingsBoard platform and was designed specifically for operating distributed 
 sensor networks for scientific research.
-Originally developed for and validated in the ACROPOLIS urban CO2 sensor network, it is network-agnostic and can be used with a 
-wide range of sensor hardware.
+Originally developed for and validated in the ACROPOLIS urban CO2 sensor network, it is network-agnostic and can be used with a wide range of sensor hardware.
 TEGAT offers a stable and reusable architectural baseline for distributed sensor networks, enabling users to 
 focus on application-specific logic while relying on a field-tested software solution for communication, data 
 persistence, and remote management of sensor devices.
 It is designed to be robust against network and power outages, crashes, and other failures, thus reducing the risk of 
 data loss, system downtime, or the need for physical intervention.
 Application-specific and hardware interfacing logic is delegated to a user-provided controller software which is managed 
-and independently deployed by TEGAT, thus separating infrastructure and application logic.
+and deployed by TEGAT, thus separating infrastructure and application logic.
 
 # Statement of need
 
 Distributed sensor networks are a critical tool in scientific research and widely used across disciplines, enabling 
-long-term, continuous sensor measurements. While they vary in the sensor hardware, data acquisition and processing 
-protocols used, as well as in the number of deployed devices, such networks often face common infrastructure challenges:
-Physical access to sensor devices is often limited or costly. This can be additionally exacerbated by challenges in 
+long-term, continuous sensor measurements. While they vary in the size and type of sensor hardware and data processing
+protocols used, such networks often face common infrastructure challenges:
+Physical access to sensor devices is often limited or costly, which can be exacerbated by challenges in 
 scaling networks to large numbers of deployed sensors. Sensor networks often need to provide continuous
-measurements while operating unattended for extended periods of time. At the same time, network connectivity and system 
+measurements while operating unattended for extended periods of time, all while network connectivity and system 
 power can be intermittent or unreliable.
 Although these challenges vary across deployments, they translate into a common set of infrastructure requirements:
 
-- Reliable bidirectional communication with a central platform
+- Reliable bidirectional communication
 - Local buffering of data during network outages
 - Remote configuration and maintenance capabilities
 - Safe remotely initiated software updates
@@ -78,10 +68,9 @@ Although these challenges vary across deployments, they translate into a common 
 - Seamless addition and removal of devices from the network
 
 TEGAT addresses these infrastructure requirements to significantly reduce the engineering overhead associated 
-with deploying and maintaining sensor networks. This enables network operators to focus on application-specific logic 
-while building on top of field-tested software. TEGAT leverages the ThingsBoard IoT platform (@ThingsBoard), a 
-robust open source software, which was chosen for its maturity (10+yrs in development) and scalability, supporting large 
-numbers of sensor devices.
[…]
