### 2026-06-10 5a919e0
add references header at end of paper.md
