### 2026-07-16 4ca143e
fix(db_schemas): turn all table query strings into format strings
