### 2026-07-16 f404f15
fix(db_schemas): turn all table query strings into format strings (#31)
