### 2026-07-16 0ab96ec
fix(docs): clarify production setup vs demo docs, deduplicate docs
