### 2026-07-16 4fa1fef
Apply suggestions from code review
Co-authored-by: Lars Frölich <lars.froelich@gmail.com>
