+++ demo/README.md
@@ -1,6 +1,6 @@
 # Local Demo (quick start) for local testing
 This folder contains a demo setup for local testing of TEGAT.
-It includes a thingsboard server instance, a TEGAT instance as well as an example controller software 
+It includes a ThingsBoard server instance, a TEGAT instance as well as an example controller software 
 producing simulated data for visualization in ThingsBoard.
 
 Detailed instructions on how to run the demo can be found in the demo docs [here](https://tum-esm.github.io/TEGAT/getting-started/2_demo.html)
