### 2026-07-16 5f411bf
expanded on demo/README.md description text
