+++ demo/README.md
@@ -1,6 +1,10 @@
 # Local Demo (quick start) for local testing
-This folder contains a demo setup for local testing of TEGAT.
-It includes a ThingsBoard server instance, a TEGAT instance as well as an example controller software 
-producing simulated data for visualization in ThingsBoard.
+This folder contains the setup/code files for a demo deployment of TEGAT.
+The demo-deployment allows running TEGAT locally on a developer's machine using Docker, 
+and includes a ThingsBoard server instance, a TEGAT instance as well as an example controller software.
+This is intended for 
+- local testing of TEGAT features and functionality via the ThingsBoard platform
+- local development and testing of TEGAT's controller software
+- integration testing of TEGAT with ThingsBoard and the controller software
 
 Detailed instructions on how to run the demo can be found in the demo docs [here](https://tum-esm.github.io/TEGAT/getting-started/2_demo.html)
