### 2026-07-16 09b14f6
fix(docs): clarify production setup vs demo docs, deduplicate docs (#32)
* fix(docs): clarify production setup vs demo docs, deduplicate docs

* Apply suggestions from code review

Co-authored-by: Lars Frölich <lars.froelich@gmail.com>

* expanded on demo/README.md description text
