### 2026-08-16 #10887 @abhishektiwari
@patrickjaigner Please see suggested improvements/issued in manuscript. Please make changes, generated PDF using `@editorialbot generate pdf`, review it to ensure that formatting, reference, and other issues are resolved in generated PDF. Then I can read a final time before moving a recommendation. Let me know if any questions.

## Main Body (Edits in paper.md):

### Pandoc citation formatting issues

Every in-text citation renders as nested parentheses: "(thingsboard contributors (2025))", "(Merkel (2014))", "(Torvalds et al. (2026))", etc. This is the Pandoc `[@key]` vs. `@key` distinction: you want `[@key]`, which produces "(Merkel, 2014)". This affects lines 38, 48–49, 52–55, 59–61, 62, 77, 80, 91, 99, 116, 120, 124, 126, 132, 133.

### Typos and inconsistent proper nouns in the body

- **Line 48**: "ThingsBoard's own product **Thingsboard** Edge" should be ThingsBoard.
- **Lines 60–61**: "**Kube Edge**" should be KubeEdge (one word).
- **Line 98**: "a local **sqlite** database" should be SQLite (Figure 1 already uses SQLite).
- **Lines 104, 117**: "the Controller Software's **docker** container", "builds a **docker** image" should be Docker (capitalized correctly at line 80).
- **Line 85**: "the Controller **software**" should be Controller Software, for consistency with the rest of the paper.
- **Line 127**: "GitHub **actions**" should be GitHub Actions.
- **Line 141**: "ICOS **cities** project" should be ICOS Cities (as at line 133).
- **Line 147**: "HORIZON EUROPE" should be Horizon Europe; "consolidator grant" should be Consolidator Grant.
- **Line 17**: "a user-provided **controller software**" should be Controller Software.

### Suggested improvements

- **Line 7**: "is **a** lightweight, general-purpose open source software": *software* is uncountable; drop the article, or write "is a lightweight, general-purpose open source **tool**".
- **Line 11**: "it is **network-agnostic** and can be used with a wide range of sensor hardware". Given the second clause, you likely mean **hardware-agnostic**.
- **Line 16**: "Application-specific and hardware interfacing logic" should be "hardware-interfacing logic" (hyphenate the compound modifier).
- **Line 39**: "(10+yrs in development)" should be "(more than 10 years in development)".
- **Line 51**: "**Some** subset of our features" should be "A subset".
- **Lines 57 and 62**: "Finally," opens two sentences in the same paragraph. Change the first to "However," or "Alternatively,".
- **Line 63**: "**doesn't** separate" should be "does not separate" (the only contraction in the paper).
- **Line 72**: "with TEGAT acting as intermediary" should be "as **an** intermediary".
- **Line 74**: "divides responsibilities **between** all three components" should be **among**.
- **Lines 102, 112**: "on-demand" used adverbially should be "on demand" (no hyphen).
- **Line 126**: "Developers can perform local type checks using mypy, **which is also deployed as** a continuous integration pipeline": the antecedent is unclear. Suggest: "mypy, which also **runs in** a continuous integration pipeline via GitHub Actions."
- **Line 135**: "it continuously ran **over** 18 months" should be "for over 18 months". This paragraph uses "over" four times in two sentences; consider "more than" for some.
- **Line 139**: "LF, PA, designed the software architecture": stray comma should be "LF and PA designed".
- **Lines 144–145**: "funded by: PAUL … **funded by** the European Union's Horizon 2020": the repetition is awkward; drop the colon and recast as "This work was funded by PAUL … a project of the ICOS Cities programme under the European Union's Horizon 2020…".

Also,

- **JOSS Sty**: Your section headings mix sentence case ("Statement of need", "Research impact statement", "Author contributions") with title case ("State of the Field", "Software Architecture", "Software Design and Implementation"). JOSS papers conventionally use sentence case throughout, so I'd lowercase the latter three.

## References (Edits in paper.bib): Please use {} explicitly to capitalise/uncapitalize as by default JOSS formats reference titles as sentence case.

### Proper nouns and capitalization

Most entries use the project's own lowercase styling for the author field, which BibTeX preserves but which reads as an error in a reference list. Suggested fixes:

- **Line 162**: "azure-iotedge contributors" should be **Azure IoT Edge contributors**
- **Line 166**: "fledge contributors" should be **Fledge contributors**; title *FL Edge* should be *Fledge*
- **Line 169**: "sqlite contributors" should be **SQLite contributors**
- **Line 172**: "kapua contributors" should be **Kapua contributors**
- **Line 174**: "KOMIYA, T. … SHIMIZUKAWA, T." should be **Komiya, T. … Shimizukawa, T.**
- **Line 179**: "kubeedge contributors" should be **KubeEdge contributors**
- **Line 182**: "kura contributors" should be **Kura contributors**
- **Line 191**: "mypy contributors" should be **Mypy contributors**
- **Line 194**: "thin-edge.io contributors" should be **Thin-edge.io contributors**
- **Line 196**: "thingsboard contributors" should be **ThingsBoard contributors**
- **Line 199**: "thingsboard-edge contributors" should be **ThingsBoard Edge contributors**
- **Line 204**: "linux kernel contributors" should be **Linux kernel contributors**
- **Line 187**: "Lightweight linux containers" should be **"Lightweight Linux containers"**

### Other errors in the references

- **Line 164: "Belana cloud"** is a misspelling of **Balena Cloud**, and the URL is wrong: it points to `https://aws.amazon.com/iot`.
- **Line 160: "AWS-IoT"** should be "AWS IoT" (no hyphen).
- **Lines 179–180**: the KubeEdge title reads "Cloud Native Computing Foundation: Kubernetes Native Edge Computing Framework". The project name is missing; should be "KubeEdge: Kubernetes native edge computing framework".
- **Lines 199–200**: the ThingsBoard Edge title is duplicated: "ThingsBoard Edge: Open Source IoT Edge Computing ThingsBoard Edge: IoT that works when the internet doesn't".
- **Lines 166–168**: the Fledge title is a comma splice ("…Internet of Things, it acts as an edge gateway…"). Trim to just the project description.
- **Lines 161, 165, 190, 193**: stray periods after URLs (`…/iot.`, `…mqtt-specification/.`, `…python.org/.`)
- **Line 175: "B., J.-F."** Looks a malformed author name from the upstream Zenodo metadata. Please use correct bibtex entry with all author names.

```
@software{takeshi_komiya_2025_18109073,
  author       = {Takeshi KOMIYA and
                  Georg Brandl and
                  Adam Turner and
                  Jean-François B. and
                  Takayuki SHIMIZUKAWA and
                  Jakob Lykke Andersen and
                  Daniel Neuhäuser and
                  Stephen Finucane and
                  Robert Lehmann and
                  jacobmason and
                  Timotheus Kampik and
                  James Addison and
                  Justus Magin and
                  Jon Dufresne and
                  Jonathan Waltman and
                  danieleades and
                  Juan Luis Cano Rodríguez and
                  Armin Ronacher and
                  Matthias Geier and
                  Dmitry Shachnev and
                  Yoshiki Shibukawa and
                  Bénédikt Tran and
                  Adam Dangoor and
                  Chris Sewell and
                  Rob Ruana and
                  Pauli Virtanen and
                  Tim Hoffmann and
                  Martin Liška and
                  François Freitag},
  title        = {sphinx-doc/sphinx: Sphinx 9.1.0},
  month        = dec,
  year         = 2025,
  publisher    = {Zenodo},
  version      = {v9.1.0},
  doi          = {10.5281/zenodo.18109073},
  url          = {https://doi.org/10.5281/zenodo.18109073},
  swhid        = {swh:1:dir:7044bddc1f760402c00f5db60b6e4fef4ccc4a38
                   ;origin=https://doi.org/10.5281/zenodo.7857310;vis
                   it=swh:1:snp:1cf71a75ed8fcc9f43533c53a1527b88ba2b6
                   a45;anchor=swh:1:rel:a03384f99b76713242d3aaff9a318
                   0d868f1ed92;path=sphinx-doc-sphinx-a7a88a2
                  },
}
```
