### 2026-08-18 #10887 @patrickjaigner
Dear @abhishektiwari,
Thank you for the feedback and careful review. We appreciate the suggested changes, which have improved the overall quality and consistency of the manuscript. We have implemented all suggested changes in the merged [PR](https://github.com/tum-esm/TEGAT/pull/33), with one exception concerning “network-agnostic” in Line 11. This wording is intentional, as it refers to TEGAT being independent of the specific sensor network in which it is deployed, rather than to the sensor hardware itself. Please let us know if there is anything else we can provide or address.
