### 2026-08-24 #10887 @danielskatz
👋 @patrickjaigner - I've suggested a few minor changes in the paper in https://github.com/tum-esm/TEGAT/pull/34 - please merge this, or let me know what you disagree with
