### 2026-08-24 #10887 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
