### 2026-08-24 #10887 @danielskatz
👋 @patrickjaigner - I noticed I fixed one list, but missed a second, which I've now attempted to fix in https://github.com/tum-esm/TEGAT/pull/35 - please merge this as well, if it seems ok
