### 2026-08-24 #10887 @danielskatz
Congratulations to @patrickjaigner (Patrick Aigner) and co-authors on your JOSS publication!!

And thanks to @Puneetha-Ramachandra, @Ahanaf-Aziz & @brpat07 for reviewing, and to @abhishektiwari for editing!
JOSS completely depends on volunteers; we couldn't do this without you
